Refactoring in your general direction.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@1266 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Load_okt.cpp
@@ -428,7 +428,7 @@ bool CSoundFile::ReadOKT(FileReader &file)
 			SmpLength i = mptSample.nLength; 
 			while(i--)
 			{
-				*data = Clamp(*data * 2, int8(-128), int8(127));
+				*data = static_cast<uint8>(Clamp(static_cast<int>(*data) * 2, int8(-128), int8(127)));
 				*data++;
 			}
 		}
